Split TestSimpleOffsets into a couple files to make working on the test failures easier
Worked on fixing the regression that happened from the fixes for units checking.
What was happening was that for this:

  Dec 31st 1945

The s in st was being captured by the Units or QUnits check and causing a 31 seconds adjustment
to happen. I added a check inside of each to make sure that it wasn't that edge case.

All but one test is passing now - more later