Test failures are now more informative with reusable and less obscure timetuple comparisons.
This primarily addresses the "digging" that was necessary when a test failed. Since printed output is not in line with test results it was not only necessary to print the values that are failing, but also run only that individual test. With this change the result and target value are printed in the test failure message.